// UI/src/app/Models/Customer/StagesAndPayments.ts
export class CustomerStagesPaymentsDto {
  paymentDetailsId: number = 0;
  paidAmount: number = 0;
  paymentDate: Date = new Date();
  balanceAmount: number = 0;
  stageId: number = 0;
  paymentTypeId: number = 0;
  accountNumber: string = '';
  firstApplicantName: string = '';
  firstApplicantMiddleName: string = '';
  firstApplicantLastName: string = '';
  secondApplicantFirstName: string = '';
  secondApplicantMiddleName: string = '';
  secondApplicantLastName: string = '';
  projectName: string = '';
  blockName: string = '';
  unitName: string = '';
  unitTypeName: string = '';
  projectId: number = 0;
  blockId: number = 0;
  levelId: number = 0;
  unitId: number = 0;
  stageName: string = '';
  paymenetStatus: string = '';
  totalAmount: number = 0;
  totalTds: number = 0;
  levelName: string = '';
  applicantId: number = 0;
  referenceNumber: string = '';
  customerBankName: string = '';
  paymentMode: string = '';
  demandLetter: string = '';
  finalPrice: number = 0;
  pendingAmount: number = 0;
  paidAmountForApproval: number = 0;
  actionStatus: string = '';
  balanceAmountForApproval: number = 0;
  //	private Double totalInitiatedAmountWithoutTds;
  //	private Double totalInitiatedAmountWithTds;
  receivedAmount: number = 0;
  //	private Double totalInitiatedBalanceAmount;
  finalPriceWithoutTds: number = 0;
  paymentStatusId: number = 0;
  paymentStatus: string = '';
  paymentReceiptUrl: string = '';
  formattedPendingAmount: string = '';
  formattedPaidAmount: string = '';
  formattedPaidAmountForApproval: string = '';
  firstApplicantPhoneNumber: string = '';
  formattedFinalPrice: string = '';
  formattedFinalPriceWithoutTds: string = '';
}
export class StageAndPaymentsDto {
  demandLetter: string = '';
  stageId: number = 0;
  stageName: string = '';
  stageAmountWithTds: number = 0;
  stageAmountWithoutTds: number = 0;
  stagePendingAmount: number = 0;
  stagePaidAmount: number = 0;
  stageTds: number = 0;
  formattedStageAmountWithTds: string = '';
  formattedStageAmountWithOutTds: string = '';
  formattedStagePendingAmount: String = '';
  formattedStageTds: String = '';
  formattedStagePaidAmount: string = '';
  paymentDto: CustomerStagesPaymentsDto[] = [];

  // Array of PaymentDetailsDto objects
}

export class CustomerUnitPaymentDetailsDto {
  projectName: string = '';
  unitName: string = '';
  levelName: string = '';
  finalPriceWithoutTds: number = 0;
  finalPriceWithTds: number = 0;
  totalInitiatedAmountWithTds: number = 0;
  totalInitiatedAmountWithoutTds: number = 0;
  totalPaidAmount: number = 0;
  totalBalanceAmount: number = 0;
  applicantId: number = 0;
  stagesDto: StageAndPaymentsDto[] = [];
  formattedFinalPriceWithoutTds: string = '';
  formattedFinalPriceWithTds: string = '';
  formattedTotalInitiatedAmountWithTds: string = '';
  formattedTotalInitiatedAmountWithoutTds: string = '';
  formattedTotalPaidAmount: String = '';
  formattedTotalBalanceAmount: string = '';
}